pyproject.toml
[tool.poetry]
name = "ghost-ids"
version = "0.1.0"
description = ""
authors = ["mikolaj-skrzypczak <skrzypczak666@gmail.com >"]
packages = [
    { include = "ids_filler" }
]

[tool.poetry.scripts]
prepare_ids = "ids_filler.__main__:main"

[tool.poetry.dependencies]
python = "3.10"
Pillow = "^9.4.0"
PyYAML = "^6.0"
aspose-words = "^23.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
